Safari-611.1.10/LayoutTests/fast/text/font-weight-download.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpResourceLoadCallbacks();
    testRunner.dumpAsText();
}
if (window.internals) {
    internals.invalidateFontCache();
    internals.clearMemoryCache();
}
</script>
<style>
@font-face {
    font-family: "WebFont";
    font-weight: 100;
    src: url("../../resources/Ahem.ttf") format("truetype");
}
@font-face {
    font-family: "WebFont";
    font-weight: 900;
    src: url("../../resources/Ahem.otf") format("opentype");
}
</style>
</head>
<body>
This test makes sure that unnecessary fonts aren't downloaded. The test fails if Ahem.otf is downloaded.
<div style="font: 100 100px 'WebFont';">Hello</div>
</body>
</html>
